+++ src/compiler/generators/wrappers/DOMComponentWrapper.js
@@ -4,6 +4,7 @@ import __name__ from "__import__";
 const app = new __name__({
   target: __target__,
   props: __data__,
+  hydrate: __hydrate__,
 });
 
 export default app;