+++ types/global.d.ts
@@ -2,5 +2,7 @@ declare global {
   const __data__: unknown;
   const __target__: string;
   const __name__: string;
+  const __hydrate__: boolean;
+  const __context__: boolean;
 }
 export {};