+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 ### Bug Fix
 
+### Features
+
+- Added new bundle type, `hydrate`. It will create a hydratable bundle that combines `ssr` and `dom`.
+
+- Added README.md.
+- Added `examples` section.
+
+## v0.0.2
+
+### Bug Fix
+
 - Fix absolute working dir for esbuild
 
 ### Features